Merges the add-burritocoin-tests-1mPTk branch which contains:
- BurritoCoin genesis block (mainnet/testnet/regtest) with WSJ 18/Mar/2026
  headline and team-controlled pubkey 04dd6fb3...
- VPS testnet/mainnet config files and setup script
- btc-rpc-explorer integration for block explorer
- 30+ additional bug fixes: assert → proper error handling throughout,
  loop bounds underflow protection (i+1 < size), pindexPrev rename,
  GetBalance MWEB integration, RemoveWallet deadlock fix, null pointer
  guards, CFeeRate rename (nSatoshisPerK → nBurrioshisPerK), etc.

Conflict resolution (5 files):
- tx_check.cpp: identical code in both branches; kept theirs (adds comment)
- mweb_miner.cpp: kept our detailed error message; fixed pIndexPrev typo
- mweb_node.cpp: took their proper BlockValidationResult error return
  over our assert (superior — returns meaningful error code to caller)
- wallet.cpp: took their early-return for unmatched pegout
  (size invariant check below handles the invariant separately)
- regtest.sh: took their version (5 iterations of fixes applied)

Our structural IsHogEx() fix (removing m_hogEx flag) is preserved and
auto-merged correctly into transaction.h and transaction.cpp.

+++ src/consensus/params.h
@@ -15,7 +15,7 @@ enum DeploymentPos
 {
     DEPLOYMENT_TESTDUMMY,
     DEPLOYMENT_TAPROOT, // Deployment of Schnorr/Taproot (BIPs 340-342)
-    DEPLOYMENT_MWEB, // Deployment of MWEB (LIPs 0002-0004)
+    DEPLOYMENT_MWEB, // Deployment of MWEB (BIPs 0002-0004)
     // NOTE: Also add new deployments to VersionBitsDeploymentInfo in versionbits.cpp
     MAX_VERSION_BITS_DEPLOYMENTS
 };
@@ -55,7 +55,7 @@ struct BIP9Deployment {
  */
 struct Params {
     uint256 hashGenesisBlock;
-    int nSubsidyHalvingInterval;
+    int64_t nSubsidyHalvingInterval;
     /** Block height at which BIP16 becomes active */
     int BIP16Height;
     /** Block height and hash at which BIP34 becomes active */
